flask_test/requirements.txt
# This file defines the image's dependencies
Flask
Flask-RESTful
google-cloud-datastore==1.7.3
pyzipcode3
geopy
flask_restful
geocoder
Jinja2
Werkzeug
flask_admin
sqlalchemy
stripe